Generating queryresults using our system

+++ rxs6616-assignment2/src/main/java/edu/louisiana/cacs/csce561/assignment2/model/QueryResult.java
@@ -1,6 +1,5 @@
 package edu.louisiana.cacs.csce561.assignment2.model;
 
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
@@ -8,7 +7,7 @@ public class QueryResult {
 	
 	private Query query;
 	
-	private Map<String,Integer> fileRSVMap = new LinkedHashMap<String, Integer>();
+	private Map<String,Double> fileRSVMap = new LinkedHashMap<String, Double>();
 
 	public Query getQuery() {
 		return query;
@@ -18,11 +17,11 @@ public class QueryResult {
 		this.query = query;
 	}
 
-	public Map<String, Integer> getFileRSVMap() {
+	public Map<String, Double> getFileRSVMap() {
 		return fileRSVMap;
 	}
 
-	public void setFileRSVMap(Map<String, Integer> fileRSVMap) {
+	public void setFileRSVMap(Map<String, Double> fileRSVMap) {
 		this.fileRSVMap = fileRSVMap;
 	}
 }
